+++ src/main/java/bowling/game/frame/FinalPitches.java
@@ -1,4 +1,4 @@
-package bowling.game;
+package bowling.game.frame;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -6,18 +6,19 @@ import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.joining;
 
-public class NormalPitches implements Pitches {
+public class FinalPitches implements Pitches {
     private static final int PITCH_COUNT_MAX = 2;
+    private static final int BONUS_COUNT_MAX = 3;
 
     private final LinkedList<Pitch> pitches;
 
-    public NormalPitches() {
+    public FinalPitches() {
         this.pitches = new LinkedList<>();
     }
 
     @Override
     public void throwBall(final int pinCount) {
-        if (pitches.isEmpty()) {
+        if (pitches.isEmpty() || isBonusPitch()) {
             pitches.add(Pitch.firstPitch(pinCount));
             return;
         }
@@ -28,17 +29,11 @@ public class NormalPitches implements Pitches {
 
     @Override
     public boolean hasChance() {
-        if (pitches.isEmpty()) {
-            return true;
-        }
-
-        Pitch firstPitch = pitches.getFirst();
-
-        if (firstPitch.isStrikePitch()) {
+        if (pitches.size() == PITCH_COUNT_MAX && !isBonusPitch()) {
             return false;
         }
 
-        if (pitches.size() == PITCH_COUNT_MAX) {
+        if (pitches.size() == BONUS_COUNT_MAX) {
             return false;
         }
 
@@ -50,6 +45,10 @@ public class NormalPitches implements Pitches {
         return pitches.stream()
                 .map(Pitch::stateToString)
                 .collect(joining("|"));
+    }
+
+    private boolean isBonusPitch() {
+        return pitches.getFirst().isStrikePitch() || pitches.getLast().isSparePitch();
     }
 
     @Override
@@ -64,15 +63,12 @@ public class NormalPitches implements Pitches {
 
     @Override
     public int getBasicScore() {
-        if (this.isStrikePitches() || this.isSparePitches()) {
-            return 10;
-        }
-
         return pitches.stream()
                 .mapToInt(Pitch::getPinCount)
                 .sum();
     }
 
+    @Override
     public List<Integer> getPitchesPinCounts() {
         return pitches.stream()
                 .map(Pitch::getPinCount)
