+++ src/test/java/bowling/game/frame/FrameNumberTest.java
@@ -1,12 +1,11 @@
-package bowling.game;
+package bowling.game.frame;
 
+import bowling.game.frame.FrameNumber;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class FrameNumberTest {
