+++ src/test/java/bowling/game/frame/FramesTest.java
@@ -1,9 +1,10 @@
-package bowling.game;
+package bowling.game.frame;
 
+import bowling.game.Score;
+import bowling.game.frame.Frame;
+import bowling.game.frame.Frames;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
