+++ src/test/java/bowling/game/frame/NormalPitchesTest.java
@@ -1,5 +1,8 @@
-package bowling.game;
+package bowling.game.frame;
 
+import bowling.game.frame.FinalPitches;
+import bowling.game.frame.NormalPitches;
+import bowling.game.frame.Pitches;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -9,7 +12,6 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NormalPitchesTest {
