+++ src/test/java/bowling/game/frame/PitchTest.java
@@ -1,7 +1,8 @@
-package bowling.game;
+package bowling.game.frame;
 
+import bowling.game.State;
+import bowling.game.frame.Pitch;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
